Cleanup Images and added build scripts
There were duplicates in Diagrams and Images but the ones from Diagrams
were references. The new pipeline expects images in the Images folder though.
Since the files were identical, this commit removes the png files from
the Diagrams folder and makes sure that we are using the ones from Images.

This commit also adds the basic build and setup scripts for the new
workflow.